requirements.txt
streamlit==1.52.1
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.8.0
xgboost==3.2.0
lightgbm==4.6.0
catboost==1.2.10
joblib==1.5.3
plotly==6.5.0
matplotlib==3.10.8
seaborn==0.13.2
shap==0.51.0
requests==2.32.5
markdown==3.10.2
python-dotenv==1.2.1 
google-genai==1.67.0 
fastapi==0.135.1 
uvicorn==0.41.0
ipython==9.8.0
requests==2.32.5